refactor: 컷아웃 에러 처리 및 업로드 상태 관리를 useImageUploadAndCutout으로 이동
